:wrench: Fix gcc compile errors

+++ src/child.c
@@ -2,8 +2,9 @@
 
 bool child_inc(child_t *self, size_t order)
 {
-	size_t num_pages = 1 << order;
-	if (self->huge == true || self->free + num_pages > LLFREE_CHILD_SIZE)
+	uint16_t num_pages = (uint16_t)(1u << order);
+	if (self->huge == true ||
+	    (uint16_t)(self->free + num_pages) > LLFREE_CHILD_SIZE)
 		return false;
 
 	self->free += num_pages;
@@ -12,7 +13,7 @@ bool child_inc(child_t *self, size_t order)
 
 bool child_dec(child_t *self, size_t order, bool allow_reclaimed)
 {
-	size_t num_pages = 1 << order;
+	uint16_t num_pages = (uint16_t)(1u << order);
 	if (!self->huge && self->free >= num_pages &&
 	    (allow_reclaimed || !self->reclaimed)) {
 		self->free -= num_pages;
